system/mcperf: explicitly set server to 127.0.0.1
The memcached server listens on 127.0.0.1 by default on some Linux
distros such as Ubuntu 20.04, while mcperf connects to localhost if the
server parameter is not provided. In some cases localhost may resolve to
the ipv6 address instead of 127.0.0.1 and result in a connection error.
This change set the server to 127.0.0.1 explicitly.